Feature/aos/main feed design

See merge request s04-final/s04p31b206!123

+++ AOS/app/src/main/java/com/omnyom/yumyum/interfaces/RetrofitService.kt
@@ -65,7 +65,7 @@ interface RetrofitService {
 
     // 피드 좋아요!
     @POST("feed/like")
-    fun feedLike(@Body parameters: HashMap<String, Int>) : Call<LikeResponse>
+    fun feedLike(@Body parameters: HashMap<String, Long>) : Call<LikeResponse>
 
     // 피드 좋아요 취소!
     @DELETE("feed/like/{feedId}/{userId}")
